create MainIpc_WorkspaceHandlers
